# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Red Light Green Light Platform API",
    "version": "1.0.0",
    "description": "Unified public API served through the gateway. Generated from per-service controller annotations — do not edit by hand (run `npm run generate:openapi`).",
    "contact": {
      "name": "Red Light Green Light Support"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "API Gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateProjectRequest": {
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["name"],
        "type": "object"
      },
...
      "Project": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TestRun": {
        "properties": {
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "projectId": {
            "type": "integer"
          },
          "status": {
            "enum": ["open", "closed"],
            "type": "string"
          }
        },
        "type": "object"
      },
      "TestRunInput": {
        "properties": {
          "caseIds": {
            "description": "One or more test case IDs to include in the run",
            "items": {
              "type": "integer"
            },
            "type": "array"
          },
          "description": {
            "maxLength": 500,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "type": "string"
          },
          "projectId": {
            "type": "integer"
          },
          "suiteId": {
            "type": "integer"
          }
        },
        "required": ["name", "projectId", "caseIds"],
        "type": "object"
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "List of projects"
          }
        },
        "summary": "List projects for current user",
        "tags": ["Projects"]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Project created successfully"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "summary": "Create a new project",
        "tags": ["Projects"]
      }
    },
...
    "/api/v1/testruns": {
      "get": {
        "parameters": [
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestRun"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of test runs"
          }
        },
        "summary": "List all test runs in a project",
        "tags": ["TestRuns"]
      },
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Optional key to make run creation idempotent on retry",
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRunInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRun"
                }
              }
            },
            "description": "Created test run"
          },
          "400": {
            "description": "Validation failed or invalid case IDs"
          },
          "403": {
            "description": "projectId mismatch with x-project-id header"
          }
        },
        "summary": "Create a new test run",
        "tags": ["TestRuns"]
      }
    },
...
  }
}